Relocate TableRow to root table types
