+++ javascript/packages/core/components/table/__tests__/table.test.tsx
@@ -16,10 +16,11 @@ import {
   expectTableRows,
 } from '../__fixtures__/table-test-helpers';
 import { FilterMode } from '../components/filter/types';
-import { TableBodyProps, TableRow } from '../components/table-body/types';
+import { TableBodyProps } from '../components/table-body/types';
 import { useTableSelectionContext } from '../plugins/selection/table-selection-context';
 import { Table } from '../table';
 
+import type { TableRow } from '#core/components/table/types/row-types';
 import type { Accessor } from '#core/types/common/studio-types';
 import type { TablePaginationProps } from '../components/table-pagination/types';
