+++ javascript/packages/core/components/table/components/table-body/row-transformer.ts
@@ -5,7 +5,7 @@ import { TableCellContent } from './table-cell-content';
 import type { Row } from '@tanstack/react-table';
 import type { ColumnConfig } from '#core/components/table/types/column-types';
 import type { TableData } from '#core/components/table/types/data-types';
-import type { TableRow } from './types';
+import type { TableRow } from '#core/components/table/types/row-types';
 
 export function transformRows<T extends TableData = TableData>(
   tanstackRows: Row<T>[]
