+++ javascript/packages/core/components/table/components/table-body/types.ts
@@ -1,23 +1,6 @@
-import type { ReactNode } from 'react';
-import type { SelectableCapability } from '#core/components/table/types/column-types';
-import type { ColumnConfig } from '#core/components/table/types/column-types';
 import type { TableData } from '#core/components/table/types/data-types';
+import type { TableRow } from '#core/components/table/types/row-types';
 import type { WithStickySidesProps } from '../with-sticky-sides/types';
-
-export type TableCell<T extends TableData = TableData> = {
-  id: string;
-  content: ReactNode;
-  column: ColumnConfig<T>;
-  value: unknown;
-  isVisible: boolean;
-};
-
-export type TableRow<T extends TableData = TableData> = {
-  id: string;
-  cells: TableCell<T>[];
-  record: T;
-} & SelectableCapability &
-  ExpandableCapability;
 
 export type TableBodyProps<T extends TableData = TableData> = {
   rows: TableRow<T>[];
@@ -25,20 +8,3 @@ export type TableBodyProps<T extends TableData = TableData> = {
   subRow?: React.ComponentType<{ row: TableRow<T> }>;
   actions?: React.ComponentType<{ row: TableRow<T> }>;
 } & Pick<WithStickySidesProps, 'enableStickySides' | 'scrollRatio'>;
-
-/**
- * Defines a row's expansion capabilities and current state.
- * Used by expandable row components to enable sub-row expansion interactions.
- *
- * @example
- * ```ts
- * // isExpanded = false
- * onToggleExpanded()
- * expect(isExpanded).toBe(true)
- * ```
- */
-export type ExpandableCapability = {
-  canExpand: boolean;
-  isExpanded: boolean;
-  onToggleExpanded: () => void;
-};
