+++ javascript/packages/core/components/table/types/table-types.ts
@@ -2,11 +2,12 @@ import { EmptyState } from '../components/table-empty-state/types';
 
 import type { ApplicationError } from '#core/types/error-types';
 import type { TableActionBarConfig } from '../components/table-action-bar/types';
-import type { TableBodyProps, TableRow } from '../components/table-body/types';
+import type { TableBodyProps } from '../components/table-body/types';
 import type { TableHeaderProps } from '../components/table-header/types';
 import type { PageSizeOption, TablePaginationProps } from '../components/table-pagination/types';
 import type { ColumnConfig } from './column-types';
 import type { TableData } from './data-types';
+import type { TableRow } from './row-types';
 
 /**
  * @interface TableRequiredUserProps
